+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2020-02-06
+### Added
+- Allow to provide JSON response as python values.
+
+### Fixed
+- method can now be provided even if not entirely upper cased.
+
 ## [0.0.2] - 2020-02-06
 ### Added
 - Allow to retrieve requests.
@@ -14,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First release, should be considered as unstable for now as design might change.
 
-[Unreleased]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/Colin-b/pytest_httpx/releases/tag/v0.0.1